Relocate `CreateProfileCommand` `ListProfilesCommand` `ProfileInfoCommand` from `org.grails.cli.commands` to `org.grails.cli.command.profile`
