+++ grace-shell/src/main/groovy/org/grails/cli/command/profile/ProfileInfoCommand.groovy
@@ -13,12 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.commands
+package org.grails.cli.command.profile
 
 import groovy.transform.CompileStatic
 
 import grails.build.logging.GrailsConsole
 
+import org.grails.cli.commands.ArgumentCompletingCommand
+import org.grails.cli.commands.Command
+import org.grails.cli.commands.CommandDescription
+import org.grails.cli.commands.ExecutionContext
+import org.grails.cli.commands.GlobalCommand
+import org.grails.cli.commands.ProjectCommand
 import org.grails.cli.profile.Feature
 import org.grails.cli.profile.Profile
 import org.grails.cli.profile.ProfileRepository
